made shit and stuf

+++ src/Point.java
@@ -1,17 +1,17 @@
 public class Point {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
     public void printer() {
